requirements.txt
matplotlib==3.0.2
notebook==6.4.12
numpy==1.22.0
pandas==0.23.4
Pillow>=7.1.0
python-dateutil==2.7.5
python-decouple==3.1
python-engineio==3.11.1
python-slugify==3.0.4
python-socketio==2.0.0
scikit-learn==0.21.2
torch==1.13.1
torchvision==0.3.0
